+++ packages/sak-meldinger/index.ts
@@ -1,3 +1,3 @@
 export { default } from './src/MeldingerSakIndex';
 export { default as MessagesModalSakIndex } from './src/MessagesModalSakIndex';
-export type { FormValues } from './src/components/MessagesMedMedisinskeTypeBrevmal';
+export type { FormValues } from './src/components/Messages';
